+++ PIR/src/passes/removeUnusedStmts.cpp
@@ -178,7 +178,7 @@ namespace pcit::pir::passes{
 			case Expr::Kind::PARAM_EXPR:       return false;
 
 			case Expr::Kind::CALL: {
-				// TODO(FUTURE): remove if func has no side-effects
+				// TODO(FEATURE): remove if func has no side-effects
 				// if(remove_unused_stmt(stmt)){ return true; }
 
 				const Call& call_inst = handler.getCall(stmt);